chore(tree): move persistence_state out of mod.rs
